Add file based github event handling Refactor to allow resuse of event hadling in both webhook and CLI/file based triggering (GitHub Actions)
